Extract `AccessToken` parent class
This DRYs up the carrier-specific access token classes.
