+++ lib/friendly_shipping/services/tforce_freight/access_token.rb
@@ -7,35 +7,20 @@ module FriendlyShipping
     class TForceFreight
       # Represents an access token returned by TForce Freight. The access token can be
       # used to make API requests. Once it expires, a new token must be created.
-      class AccessToken
+      class AccessToken < FriendlyShipping::AccessToken
         # @return [String] the token's type
         attr_reader :token_type
-
-        # @return [Integer] the token's expiration
-        attr_reader :expires_in
 
         # @return [Integer] the token's extended expiration
         attr_reader :ext_expires_in
 
-        # @return [String] the raw JWT token
-        attr_reader :raw_token
-
         # @param token_type [String] the token's type (typically "Bearer")
-        # @param expires_in [Integer] the token's expiration
         # @param ext_expires_in [Integer] the token's extended expiration (only applicable during a service outage)
         # @see https://github.com/AzureAD/azure-activedirectory-library-for-android/issues/675 Service Outage Resiliency Support
-        # @param raw_token [String] the raw JWT token
-        def initialize(token_type:, expires_in:, ext_expires_in:, raw_token:)
+        def initialize(token_type:, ext_expires_in:, **other_kwargs)
           @token_type = token_type
-          @expires_in = expires_in
           @ext_expires_in = ext_expires_in
-          @raw_token = raw_token
-        end
-
-        # Decodes and returns the raw JWT token.
-        # @return [Array<Hash>] the decoded token
-        def decoded_token
-          @_decoded_token = JWT.decode(raw_token, nil, false)
+          super(**other_kwargs)
         end
       end
     end
